+++ src/mnist.rs
@@ -23,7 +23,7 @@ impl MnistLabels {
         Some(MnistLabels { labels })
     }
 
-    pub fn to_data(&self) -> Array2<f64> {
+    pub fn to_data(&self) -> Array2<f32> {
         let mut res = Array::zeros((self.labels.len(), 10));
         for i in 0..self.labels.len() {
             res[[i, self.labels[i] as usize]] = 1.;
@@ -66,10 +66,10 @@ impl MnistImages {
         })
     }
 
-    pub fn to_data(&self) -> Array2<f64> {
+    pub fn to_data(&self) -> Array2<f32> {
         self.images
             .to_shared()
             .reshape((self.images.len_of(Axis(0)), self.width * self.height))
-            .mapv(|x| x as f64 / 255.)
+            .mapv(|x| x as f32 / 255.)
     }
 }
